:truck: refactor: move DbAuthentication protocols to a single file
