+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:54759/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:54887/jest?","mongoDBName":"jest"}
