+++ src/data/use-cases/authentication/db-authentication-protocols.ts
@@ -1,0 +1,6 @@
+export * from '../../protocols/criptography/hash-comparer'
+export * from '../../protocols/criptography/token-generator'
+export * from '../../protocols/db/load-account-by-email-repository'
+export * from '../../protocols/db/update-access-token-repository'
+export * from '../../../domain/models/account'
+export * from '../../../domain/use-cases/authentication'
